add code for solicitud afiliacion y asignacion automatica entidad

+++ src/main/java/com/spring/security/medi/care/app/afiliacion/service/SolicitudAfiliacionService.java
@@ -10,7 +10,7 @@ public interface SolicitudAfiliacionService {
 
     SolicitudAfiliacion regristarSolicitudAfiliacion(SolicitudAfiliacion solicitud) throws Exception;
 
-    List<SolicitudAfiliacion> buscarSolicitudesAfiliacionPorParametros(String cedula,Integer servicioId, Long seguroId, Long regimenId, String estado, int page, int size);
+    List<SolicitudAfiliacion> buscarSolicitudesAfiliacionPorParametros(String cedula,Integer servicioId, Long seguroId, Integer regimenId, String estado, int page, int size);
 
     void procesarSolicitudAfiliacion(Long solicitudId) throws Exception;
